+++ src/components/week-view/week-view.js
@@ -948,13 +948,25 @@ WeekView.prototype = {
     const endMonth = Locale.formatDate(endDate, { pattern: 'MMMM', locale: this.locale.name });
     const startYear = Locale.formatDate(startDate, { pattern: 'yyyy', locale: this.locale.name });
     const endYear = Locale.formatDate(endDate, { pattern: 'yyyy', locale: this.locale.name });
-    let monthStr = Locale.formatDate(endDate, { date: 'year', locale: this.locale.name });
+    let monthStr = breakpoints.isBelow('slim') && this.currentCalendar.months ?
+      `${this.currentCalendar.months.abbreviated[endDate.getMonth()]} ${endYear}` :
+      Locale.formatDate(endDate, { date: 'year', locale: this.locale.name });
 
     if (endMonth !== startMonth) {
-      monthStr = `${Locale.formatDate(startDate, { pattern: 'MMM', locale: this.locale.name })} - ${Locale.formatDate(endDate, { pattern: 'MMMM yyyy', locale: this.locale.name })}`;
+      const startStr = breakpoints.isBelow('slim') && this.currentCalendar.months ? this.currentCalendar.months.abbreviated[startDate.getMonth()] : Locale.formatDate(startDate, { pattern: 'MMM', locale: this.locale.name });
+      const endStr = breakpoints.isBelow('slim') && this.currentCalendar.months ?
+        `${this.currentCalendar.months.abbreviated[endDate.getMonth()]} ${endYear}` :
+        Locale.formatDate(endDate, { pattern: 'MMMM yyyy', locale: this.locale.name });
+      monthStr = `${startStr} - ${endStr}`;
     }
     if (endYear !== startYear) {
-      monthStr = `${Locale.formatDate(startDate, { pattern: 'MMM yyyy', locale: this.locale.name })} - ${Locale.formatDate(endDate, { pattern: 'MMM yyyy', locale: this.locale.name })}`;
+      const startStr = breakpoints.isBelow('slim') && this.currentCalendar.months ?
+        `${this.currentCalendar.months.abbreviated[startDate.getMonth()]} ${startYear}` :
+        Locale.formatDate(startDate, { pattern: 'MMM yyyy', locale: this.locale.name });
+      const endStr = breakpoints.isBelow('slim') && this.currentCalendar.months ?
+        `${this.currentCalendar.months.abbreviated[endDate.getMonth()]} ${endYear}` :
+        Locale.formatDate(endDate, { pattern: 'MMM yyyy', locale: this.locale.name });
+      monthStr = `${startStr} - ${endStr}`;
     }
 
     this.monthField.text(monthStr);
@@ -1219,7 +1231,7 @@ WeekView.prototype = {
     }
 
     if (this.isDayView) {
-      this.currentDay = this.settings.startDate; 
+      this.currentDay = this.settings.startDate;
     }
 
     this.showWeek(this.settings.startDate, this.settings.endDate);